added feature to stay in chats while banned * ChatActivity.java do not destroy chat frame, but show popup with error (why have no access) * MessagesController.java here we allow to delete chats in DB only if we are decided to delete them, thats why method was changed * ChatObject.java   only `null` or `left==true` will return success for `isNotInChat`. `null` - thats simple, but `left` is when you left a group by yourself and deleted it. w/o `left` it will have ugly bug when you have chat UI active can type and send messages but messages are all have errors due you're not in chat. * MessagesController.java   here is simple, load all the messages.

+++ TMessagesProj/src/main/java/org/telegram/messenger/MessagesController.java
@@ -3523,10 +3523,10 @@ public class MessagesController extends BaseController implements NotificationCe
                     }
                 });
             } else {
-                AndroidUtilities.runOnUIThread(() -> {
-                    checkChannelError(error.text, chatId);
-                    loadingFullChats.remove(chatId);
-                });
+//                AndroidUtilities.runOnUIThread(() -> {
+//                    checkChannelError(error.text, chatId);
+//                    loadingFullChats.remove(chatId);
+//                });
             }
         });
         if (classGuid != 0) {
@@ -5007,6 +5007,11 @@ public class MessagesController extends BaseController implements NotificationCe
     }
 
     protected void deleteDialog(long did, int first, int onlyHistory, int max_id, boolean revoke, TLRPC.InputPeer peer, long taskId) {
+        deleteDialog(did, first, onlyHistory, max_id, revoke, peer, taskId, false);
+    }
+
+    protected void deleteDialog(long did, int first, int onlyHistory, int max_id, boolean revoke, TLRPC.InputPeer peer, long taskId, boolean isSelf) {
+        if (!isSelf) return;
         if (onlyHistory == 2) {
             getMessagesStorage().deleteDialog(did, onlyHistory);
             return;
@@ -7899,13 +7904,14 @@ public class MessagesController extends BaseController implements NotificationCe
                 AndroidUtilities.runOnUIThread(() -> checkChatInviter(chat.id, true));
             }
 
-            TLRPC.Message lastMessageFinal = lastMessage;
+//            TLRPC.Message lastMessageFinal = lastMessage;
             AndroidUtilities.runOnUIThread(() -> {
-                if (lastMessageFinal != null) {
-                    dialogsLoadedTillDate = Math.min(dialogsLoadedTillDate, lastMessageFinal.date);
-                } else {
-                    dialogsLoadedTillDate = Integer.MIN_VALUE;
-                }
+//                if (lastMessageFinal != null) {
+//                    dialogsLoadedTillDate = Math.min(dialogsLoadedTillDate, lastMessageFinal.date);
+//                } else {
+//                    dialogsLoadedTillDate = Integer.MIN_VALUE;
+//                }
+                dialogsLoadedTillDate = Integer.MIN_VALUE;
                 if (loadType != DIALOGS_LOAD_TYPE_CACHE) {
                     applyDialogsNotificationsSettings(dialogsRes.dialogs);
                     getMediaDataController().loadDraftsIfNeed();
@@ -9524,7 +9530,8 @@ public class MessagesController extends BaseController implements NotificationCe
             request = req;
         }
         if (UserObject.isUserSelf(user)) {
-            deleteDialog(-chatId, 0, revoke);
+//            deleteDialog(-chatId, 0, revoke);
+            deleteDialog(-chatId, 1, 0, 0, revoke, null, 0, true);
         }
         getConnectionsManager().sendRequest(request, (response, error) -> {
             if (error != null) {
